% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_EFTs_custering.R
\name{EFT_clust()}
\alias{EFT_clust()}
\alias{EFT_clust}
\title{EFT_clust}
\usage{
EFT_clust(obj2clust = NULL, n_clust = 20, filename = "", ...)
}
\arguments{
\item{obj2clust}{RasterStack or RasterBrick object (or its file name). Each layer is one variable}

\item{n_clust}{Numeric. Number of total clusters. Optional. Default = 20}

\item{filename}{Character. Output filename. Optional}

\item{...}{Arguments for kmeans(). Optional}
}
\value{
RasterLayer object
}
\description{
EFT_clust derives the Ecosystem Functional Types using K-means to perform a clustering on the
pixels of the Raster* object
}
\details{
kmeans() does not optimize the final number of clusters. It needs to be set by means of n_clust
(default = 20). There are several methods and statistics to determine the optimal number. clust_optim()
produces a scree plot to help the user to decide the optimal number of clusters. In this case, kmeans()
uses as default iter.max = 500 and algorithm = "MacQueen", but these can be modified passing the arguments
throuh ...
}
\examples{
\dontrun{
LPDynR:::EFT_clust(obj2clust = EFTs_raster,
                   n_clust = 20)
}

}
\author{
Xavier Rotllan-Puig
}
