% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_CombinedAssessment.R
\name{LPD_CombAssess}
\alias{LPD_CombAssess}
\title{LPD_CombAssess}
\usage{
LPD_CombAssess(LandProd_change = NULL, LandProd_current = NULL,
  filename = "")
}
\arguments{
\item{LandProd_change}{RasterLayer object (or its file name). Land Productivity Long Term Change Map}

\item{LandProd_current}{RasterLayer object (or its file name). Land Productivity Current Status Map}

\item{filename}{Character. Output filename. Optional}
}
\value{
RasterLayer
}
\description{
LPD_CombAssess combines a 'LandProd_change' map (RasterLayer) with a 'LandProd_current'
map (RasterLayer), giving a 6-classes map ranging from declining to strongly increasing land
productivity. 'LandProd_current' is reclassified into two classes: pixels with less than 50% of the
highest local production (within the EFT) and pixels with more or equal to their 50%
}
\details{
The Land Productivity Dynamics (LPD) is a qualitative indicator produced by the combined
assessment of the Land Productivity Long Term Change Map and the Land Productivity Current
Status Map
}
\examples{
\dontrun{
LPD_CombAssess(LandProd_change = LandProd_change_raster,
               LandProd_current = LandProd_current_raster)
}

}
\seealso{
\code{\link{LongTermChange}}; \code{\link{LNScaling}}
}
\author{
Xavier Rotllan-Puig
}
