% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbd_cat}
\alias{sbd_cat}
\title{Season Beginning Day}
\format{GeoTIFF. RasterBrick object with dimensions: 120, 120, 14400, 17  (nrow, ncol, ncell, nlayers)
\describe{
  \item{sbd_cat}{Season beginning day}
}}
\source{
\url{https://www.eea.europa.eu/data-and-maps}
}
\usage{
sbd_cat
}
\description{
RasterBrick object containing time series of phenological data (14400 cells; 17 layers).
Years 2000-2016. The variable has been derived from MODIS using Timesat
}
\details{
Downloaded from www.eea.europa.eu (20/08/2020)

Coord. ref. : +proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000 +ellps=GRS80 +units=m +no_defs

Resolution  : 500, 500  (x, y)

Extent      : 3620000, 3680000, 2110000, 2170000  (xmin, xmax, ymin, ymax)
}
\references{
\url{https://www.eea.europa.eu/data-and-maps/data/annual-start-of-vegetation-growing}
}
\keyword{datasets}
