% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sb_cat}
\alias{sb_cat}
\title{Standing Biomass}
\format{GeoTIFF. RasterBrick object with dimensions: 20, 20, 400, 10  (nrow, ncol, ncell, nlayers)
\describe{
  \item{sb_cat}{Standing biomass}
}}
\source{
\url{https://www.eea.europa.eu/data-and-maps}
}
\usage{
sb_cat
}
\description{
RasterBrick object containing time series of land productivity data (400 cells; 10 layers).
Years 2000-2009. The variable has been derived from MODIS using Timesat
}
\details{
Downloaded from www.eea.europa.eu (20/08/2020)

Coord. ref. : +proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000 +ellps=GRS80 +units=m +no_defs

Resolution  : 500, 500  (x, y)

Extent      : 3640000, 3650000, 2140000, 2150000  (xmin, xmax, ymin, ymax)
}
\examples{
raster::brick(paste0(system.file(package='LPDynR'), "/extdata/sb_cat.tif"))

}
\references{
\url{https://www.eea.europa.eu/data-and-maps/data/annual-above-ground-vegetation-productivity}
}
\keyword{datasets}
